more robust
first frame selection
bug fixes
