+++ Assets/Scripts/CutsceneSystem/Cutscene.cs
@@ -1,20 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 using System.Collections;
-
 
 namespace CutsceneSystem
 {
     public class Cutscene : MonoBehaviour
     {
-        private Frame[] frames;
+        public int firstFrame = 0;
+        public Frame[] frames;
         private Camera cam;
 
         private void Awake()
         {
             frames = GetComponentsInChildren<Frame>();
             cam = Camera.main;
+        }
+
+        public int GetFrameCount()
+        {
+            return GetComponentsInChildren<Frame>().Length;
         }
 
         private void Start()
@@ -24,7 +27,9 @@ namespace CutsceneSystem
 
         public IEnumerator C_StartCutscene()
         {
-            for (int frameIndex = 0; frameIndex < frames.Length; frameIndex++)
+            firstFrame = Mathf.Clamp(firstFrame, 0, frames.Length - 1);
+
+            for (int frameIndex = firstFrame; frameIndex < frames.Length; frameIndex++)
             {
                 cam.transform.position = frames[frameIndex].transform.position + new Vector3(0, 0, -10);
                 frames[frameIndex].StartFrame();
