+++ Assets/Scripts/CutsceneSystem/CutsceneEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae9dc7949cefcda4c91f6b630ad80371
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
