+++ Assets/Scripts/CutsceneSystem/Element.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections;
-using Unity.Mathematics;
+﻿using System.Collections;
 using UnityEngine;
 
 namespace CutsceneSystem
